\name{source.gist}
\alias{source.gist}
\title{
Read R code from a GitHub Gist
}
\description{
source.gist causes R to accept its input from the given GitHub Gist URL or ID. The Gist is parsed and evaluated in the global environment.
}
\usage{
source.gist(id)
}
\arguments{
  \item{id}{ a GitHub Gist URL or ID. }
}
\author{Mason Simon}

\seealso{
\code{\link{source}}
}
\keyword{ programming }
\keyword{ utilities }
\keyword{ file }
